requirements.txt
pyyaml
jsonpatch
kubernetes
rich
typer
requests
httpx
fastapi
uvicorn
numpy
pandas
scikit-learn
scipy
matplotlib
